+++ kube/Cargo.toml
@@ -17,8 +17,7 @@ edition = "2018"
 base64 = "0.11.0"
 chrono = "0.4.10"
 dirs = "2.0.2"
-serde = "1.0.104"
-serde_derive = "1.0.104"
+serde = { version = "1.0.104", features = ["derive"] }
 serde_json = "1.0.47"
 serde_yaml = "0.8.11"
 http = "0.2.0"
